break-out TypeFilterTitle
